requirements.txt
mysql-connector-python~=8.0.29
pandas~=1.4.2